Make episodes have a random unique id to avoid collisions between runs

+++ LibCarla/source/carla/client/World.cpp
@@ -17,7 +17,7 @@
 namespace carla {
 namespace client {
 
-  uint32_t World::GetId() const {
+  uint64_t World::GetId() const {
     return _episode.Lock()->GetCurrentEpisodeId();
   }
